alter content about factor levels inBar Plots section
The tutorial mentioned factor levels and dropping extra factors to get the top5 variable into alphabetical order. However, droplevels() generated an error because the top5 variable was already a character varaible rather than factored. Not sure what changed, but I altered this section to conform with what was happening in the code.